+++ packages/ice/src/commands/build.ts
@@ -1,5 +1,4 @@
 import * as path from 'path';
-import consola from 'consola';
 import fse from 'fs-extra';
 import { getWebpackConfig } from '@ice/webpack-config';
 import type { Context, TaskConfig } from 'build-scripts';
@@ -8,13 +7,14 @@ import type { StatsError, Stats } from 'webpack';
 import type { Config } from '@ice/webpack-config/esm/types';
 import type ora from '@ice/bundles/compiled/ora/index.js';
 import type { AppConfig } from '@ice/runtime/esm/types';
-import type { ServerCompiler, GetAppConfig, GetRoutesConfig, ExtendsPluginAPI } from '../types/plugin.js';
+import type { ServerCompiler, GetAppConfig, GetRoutesConfig, ExtendsPluginAPI, GetDataloaderConfig } from '../types/plugin.js';
 import webpackCompiler from '../service/webpackCompiler.js';
 import formatWebpackMessages from '../utils/formatWebpackMessages.js';
 import { IMPORT_META_RENDERER, IMPORT_META_TARGET, RUNTIME_TMP_DIR, SERVER_OUTPUT_DIR } from '../constant.js';
 import emptyDir from '../utils/emptyDir.js';
 import type { UserConfig } from '../types/userConfig.js';
 import warnOnHashRouterEnabled from '../utils/warnOnHashRouterEnabled.js';
+import { logger } from '../utils/logger.js';
 
 const build = async (
   context: Context<Config, ExtendsPluginAPI>,
@@ -25,6 +25,7 @@ const build = async (
     getAppConfig: GetAppConfig;
     appConfig: AppConfig;
     getRoutesConfig: GetRoutesConfig;
+    getDataloaderConfig: GetDataloaderConfig;
     userConfigHash: string;
     userConfig: UserConfig;
   },
@@ -36,6 +37,7 @@ const build = async (
     getAppConfig,
     appConfig,
     getRoutesConfig,
+    getDataloaderConfig,
     userConfigHash,
     userConfig,
   } = options;
@@ -65,6 +67,7 @@ const build = async (
     serverCompiler,
     getAppConfig,
     getRoutesConfig,
+    getDataloaderConfig,
   };
   const compiler = await webpackCompiler({
     context,
@@ -98,11 +101,11 @@ const build = async (
       }
 
       if (messages.errors.length) {
-        consola.error('webpack compile error');
+        logger.error('Webpack compile error.');
         reject(new Error(messages.errors.join('\n\n')));
         return;
       } else {
-        compiler?.close?.(() => {});
+        compiler?.close?.(() => { });
         const isSuccessful = !messages.errors.length;
         resolve({
           stats,
@@ -123,6 +126,7 @@ const build = async (
     serverEntryRef,
     getAppConfig,
     getRoutesConfig,
+    getDataloaderConfig,
     appConfig,
   });