+++ packages/ice/src/middlewares/mock/getConfigs.ts
@@ -2,8 +2,10 @@ import path from 'path';
 import { createRequire } from 'module';
 import assert from 'assert';
 import fg from 'fast-glob';
-import { register } from 'esbuild-register/dist/node.js';
+import esbuildRegister from '@ice/bundles/compiled/esbuild-register/node.js';
+import { logger } from '../../utils/logger.js';
 
+const { register } = esbuildRegister;
 const require = createRequire(import.meta.url);
 
 export const VALID_METHODS = [
@@ -41,8 +43,13 @@ export default function getConfigs(rootDir: string, exclude: string[] = []): Moc
   mockFiles.forEach(mockFile => {
     // disable require cache
     delete require.cache[mockFile];
-
-    const mockModule = require(mockFile);
+    let mockModule;
+    try {
+      mockModule = require(mockFile);
+    } catch (error) {
+      logger.error(`Failed to parse mock file ${mockFile}.\n${error.message}`);
+      return;
+    }
     const config = mockModule.default || mockModule || {};
     for (const key of Object.keys(config)) {
       const { method, path } = parseMockKey(key);