+++ packages/runtime/src/Document.tsx
@@ -136,7 +136,7 @@ interface DataProps {
 
 // use app context separately
 export function Data(props: DataProps) {
-  const { routesData, documentOnly, matches, routesConfig, downgrade } = useAppContext();
+  const { routesData, documentOnly, matches, routesConfig, downgrade, renderMode } = useAppContext();
   const appData = useAppData();
   const {
     ScriptElement = 'script',
@@ -152,6 +152,7 @@ export function Data(props: DataProps) {
     downgrade,
     matchedIds,
     documentOnly,
+    renderMode,
   };
 
   return (
@@ -159,7 +160,7 @@ export function Data(props: DataProps) {
     // Should merge global context when there are multiple <Data />.
     <ScriptElement
       suppressHydrationWarning={documentOnly}
-      dangerouslySetInnerHTML={{ __html: `window.__ICE_APP_CONTEXT__=Object.assign(window.__ICE_APP_CONTEXT__ || {}, ${JSON.stringify(windowContext)})` }}
+      dangerouslySetInnerHTML={{ __html: `window.__ICE_APP_CONTEXT__=Object.assign(${JSON.stringify(windowContext)}, window.__ICE_APP_CONTEXT__ || {});` }}
     />
   );
 }