+++ packages/webpack-config/src/index.ts
@@ -1,7 +1,6 @@
 import * as path from 'path';
 import { createRequire } from 'module';
 import fg from 'fast-glob';
-import consola from 'consola';
 // FIXME when prepack @pmmmwh/react-refresh-webpack-plugin
 import ReactRefreshWebpackPlugin from '@ice/bundles/compiled/@pmmmwh/react-refresh-webpack-plugin/lib/index.js';
 import bundleAnalyzer from '@ice/bundles/compiled/webpack-bundle-analyzer/index.js';
@@ -171,7 +170,7 @@ export function getWebpackConfig(options: GetWebpackConfigOptions): Configuratio
     compileExcludes,
     swcOptions,
     polyfill,
-    env: true,
+    enableEnv: true,
   });
   const webpackConfig = {
     mode,
@@ -414,7 +413,8 @@ export function getWebpackConfig(options: GetWebpackConfigOptions): Configuratio
   };
   const finalWebpackConfig = [configCss, configAssets, ...(configureWebpack || [])]
     .reduce((result, next: ModifyWebpackConfig<Configuration, typeof webpack>) => next(result, ctx), webpackConfig);
-  consola.debug('[webpack]', finalWebpackConfig);
+    // TODO: Log webpack config with namespace.
+    // consola.debug('[webpack]', finalWebpackConfig);
   return finalWebpackConfig;
 }