+++ packages/plugin-pha/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.1.0
+
+- [feat] support static dataloader set to manifest.
+
+## 1.0.3
+
+- [fix] fix lanUrlForTerminal when dev start
+- [feat] optimize log
+
 ## 1.0.2
 
 - [fix] fix dev lanUrlForTerminal err